quick and dirty solution to reversing height
Idea:

- change internal variable name to 'nite' (temporary name)
- set default value of 'height' to 1cm
- set that internal variable 'nite' to negation of 'height'

Now, when a user supplies a height value, it is negated and used exactly as height was previously used.

With this change, all the test compile correctly, although the ones that have a non-default setting for 'height' are reversed, as desired.
(except, maybe for these particular tests we want to go and change their height settings now.)

I tried to add the documentation part too, but when I try to compile the documentation pdf (running 'pdflatex braid_code.dtx') I get an error about a checksum.
I don't understand that.

I chose 'nite' for the pun: nite = n(egative) (he)ight, but spelled it 'nite' because 'height' and 'night' look really similar.
Really I just wanted a variable name that will be easy to find/replace with whatever better name someone can think of.
(E.g., 'internal height' or 'minusheight' or 'heightneg' or whatever.)

The only way I could figure out to implement this is to set a tikz key for 'nite', with initial value 1mm, then set the 'height' key with initial value 1cm, and *then* set 'nite' to be -1 times 'height'.
There must be a better way, but I don't really understand pgf/tikz well enough to do it.